+++ policy_sentry/writing/validate.py
@@ -32,9 +32,9 @@ def check(conf_schema, conf):
 CRUD_SCHEMA = Schema({
     'policy_with_crud_levels':
         {
-            'name': And(Use(str)),
-            'description': And(Use(str)),
-            'role_arn': And(Use(str)),
+            Optional('name'): And(Use(str)),
+            Optional('description'): And(Use(str)),
+            Optional('role_arn'): And(Use(str)),
             Optional('read'): [str],
             Optional('write'): [str],
             Optional('list'): [str],
@@ -48,9 +48,9 @@ CRUD_SCHEMA = Schema({
 ACTIONS_SCHEMA = Schema({
     'policy_with_actions':
         {
-            'name': And(Use(str)),
-            'description': And(Use(str)),
-            'role_arn': And(Use(str)),
+            Optional('name'): And(Use(str)),
+            Optional('description'): And(Use(str)),
+            Optional('role_arn'): And(Use(str)),
             'actions': And([str]),
         }
 })
@@ -80,3 +80,24 @@ def check_crud_schema(cfg):
         raise Exception(f"The provided template does not match the required schema for CRUD mode. "
                         f"Please use the create-template command to generate a valid YML template that "
                         f"Policy Sentry will accept.")
+
+
+def validate_condition_block(condition_block):
+    """
+    :param condition_block: {"condition_key_string": "ec2:ResourceTag/purpose", "condition_type_string": "StringEquals", "condition_value": "test"}
+    :return:
+    """
+
+    # TODO: Validate that the values are legit somehow
+    CONDITION_BLOCK_SCHEMA = Schema({
+        "condition_key_string": And(Use(str)),
+        "condition_type_string": And(Use(str)),
+        "condition_value": And(Use(str))
+    })
+    try:
+        CONDITION_BLOCK_SCHEMA.validate(condition_block)
+        # TODO: Try to validate whether or not the condition keys are legit
+        return True
+    except SchemaError as s_e:
+        print(s_e)
+        return False